Refactor package and contract
