+++ omisego-sdk/src/main/java/co/omisego/omisego/websocket/channel/SocketChannelContract.kt
@@ -11,35 +11,48 @@ import co.omisego.omisego.model.socket.SocketSend
 import co.omisego.omisego.model.socket.SocketTopic
 import co.omisego.omisego.websocket.SocketConnectionCallback
 import co.omisego.omisego.websocket.SocketListenEvent
-import co.omisego.omisego.websocket.SocketMessageRef
 import co.omisego.omisego.websocket.SocketTopicCallback
 import co.omisego.omisego.websocket.enum.SocketStatusCode
-import co.omisego.omisego.websocket.interval.SocketIntervalContract
 import okhttp3.WebSocketListener
 import java.util.Timer
 
 interface SocketChannelContract {
+    /* Channel Package */
     interface Core {
         val socketDispatcher: Dispatcher
         val socketClient: SocketClient
-        val socketMessageRef: SocketMessageRef
-        val socketHeartbeat: SocketIntervalContract
+        val socketHeartbeat: SocketInterval
+        val socketMessageRef: SocketChannelContract.MessageRef
         val heartbeatTimer: Timer?
 
         fun createJoinMessage(topic: SocketTopic, payload: Map<String, Any>): SocketSend
         fun createLeaveMessage(topic: SocketTopic, payload: Map<String, Any>): SocketSend
     }
 
+    interface MessageRef {
+        var value: String
+    }
+
+    /* WebSocket Package */
     interface SocketClient {
         fun send(message: SocketSend): Boolean
         fun closeConnection(status: SocketStatusCode, reason: String)
     }
 
+    /* Interval Package */
+    interface SocketInterval {
+        var timer: Timer?
+        val socketMessageRef: SocketMessageRef
+
+        fun startInterval(task: (SocketSend) -> Unit)
+        fun stopInterval()
+    }
+
+    /* Dispatcher Package */
     interface Dispatcher {
         fun setSocketConnectionCallback(connectionListener: SocketConnectionCallback?)
         fun setSocketTopicCallback(topicListener: SocketTopicCallback?)
         fun setSocketTransactionCallback(listener: SocketListenEvent?)
-
         fun retrieveWebSocketListener(): WebSocketListener
     }
 }
